// src/types/TenantFuncionamento.ts
export type TenantFuncionamento = {
  segOpen: string;
  segClose: string;
  terOpen: string;
  terClose: string;
  quarOpen: string;
  quarClose: string;
  quinOpen: string;
  quinClose: string;
  sexOpen: string;
  sexClose: string;
  sabOpen: string;
  sabClose: string;
  domOpen: string;
  domClose: string;
};